+++ src/test/java/com/camping/admin/helper/CommonContext.java
@@ -1,27 +1,9 @@
 package com.camping.admin.helper;
 
-import io.restassured.RestAssured;
-import io.restassured.builder.RequestSpecBuilder;
-import io.restassured.filter.log.LogDetail;
-import io.restassured.http.ContentType;
 import io.restassured.specification.RequestSpecification;
-
-import java.util.Map;
-
-import static io.restassured.RestAssured.given;
-import static io.restassured.http.ContentType.JSON;
 
 public class CommonContext {
 
     public static String adminToken;
-
-    public static RequestSpecification getRequestSpec() {
-        return new RequestSpecBuilder()
-                .setBaseUri("http://localhost")
-                .setPort(8080)
-                .setAccept(JSON)
-                .setContentType(JSON)
-                .log(LogDetail.ALL)
-                .build();
-    }
+    public static RequestSpecification requestSpec;
 }
