+++ src/test/java/com/camping/admin/helper/RequestSpecFactory.java
@@ -1,0 +1,23 @@
+package com.camping.admin.helper;
+
+import io.restassured.builder.RequestSpecBuilder;
+import io.restassured.filter.log.LogDetail;
+import io.restassured.specification.RequestSpecification;
+
+import static io.restassured.http.ContentType.JSON;
+
+public class RequestSpecFactory {
+
+    private static final String BASE_URI = "http://localhost";
+    private static final int PORT = 8080;
+
+    public static RequestSpecification getRequestSpec() {
+        return new RequestSpecBuilder()
+                .setBaseUri(BASE_URI)
+                .setPort(PORT)
+                .setAccept(JSON)
+                .setContentType(JSON)
+                .log(LogDetail.ALL)
+                .build();
+    }
+}
