+++ src/test/java/com/camping/admin/helper/TokenFactory.java
@@ -1,7 +1,4 @@
-package com.camping.admin.steps;
-
-import com.camping.admin.helper.CommonContext;
-import io.cucumber.java.BeforeAll;
+package com.camping.admin.helper;
 
 import java.util.Map;
 
@@ -13,12 +10,7 @@ import static io.restassured.http.ContentType.JSON;
  * cucumber의 BeforeAll을 사용해야함. junit BeforeAll을 사용하지 않도록 주의
  */
 @SuppressWarnings("unused")
-public class TokenHooks {
-
-    @BeforeAll
-    public static void initAdminToken() {
-        CommonContext.adminToken = getAdminToken();
-    }
+public class TokenFactory {
 
     public static String getAdminToken() {
         return given()
